ESP32/paperswarm-fallsensor/py/requirements.txt
requests
websocket-client
python-socketio
pygame
addict
